Move "createEventMergeId" into utils folder
