+++ src/utils/createEventMergeId.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { uuid } from "../../utils/index.js";
+import uuid from "./uuid.js";
 
 export default () => {
   return {
